Add initial support for setdesktopsize
When enabled with the -setdesktopsize option this allows
clients like novnc and tigervnc to request a change of
framebuffer resolution to fit that of the client window.

On the server this is implemented by leaving the physical
display resolution as-is, but setting a new framebuffer
resolution and scaling that framebuffer to the physical
display using RANDR calls.
As a result the view on the physical display may look a bit
distorted due to the scaling, but looks perfect on the remote
client. (While that is normally the opposite case)

When last client disconnects original framebuffer resolution
is restorted.

Implementation currently limited to computers with a single
screen attached.

Signed-off-by: Floris Bos <bos@je-eigen-domein.nl>